requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
pyowm
bs4
lxml
demjson
PyLyrics
wget
gTTS
#telegram
zalgo_text
pynewtonmath
bs4
telegraph
numpy
googletrans
telegraph
pillow
uuid==1.30
